+++ spec/setup/globalSetup.ts
@@ -1,9 +1,5 @@
 import '../../test-app/conf/loadEnv'
 
-import initializeDreamApp from '../../test-app/cli/helpers/initializeDreamApp'
-
-export async function setup() {
-  await initializeDreamApp()
-}
+export async function setup() {}
 
 export async function teardown() {}