+++ spec/setup/hooks.ts
@@ -4,23 +4,7 @@ import initializeDreamApp from '../../test-app/cli/helpers/initializeDreamApp'
 
 provideDreamViteMatchers(Dream)
 
-// define global context variable, setting it equal to describe
-// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-explicit-any
-;(global as any).context = describe
-
-// this is done so that the `@jest-mock/express` library can continue
-// to function. Since jest and vi have near parity, this seems to work,
-// though it is very hacky, and we should eventually back out of it.
-// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-explicit-any
-;(global as any).jest = vi
-
 beforeEach(async () => {
-  try {
-    await initializeDreamApp()
-  } catch (error) {
-    console.error(error)
-    throw error
-  }
-
+  await initializeDreamApp()
   await truncate(DreamApp)
 })