+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.5.1
+
+Capture soft-deleted records when taking snapshots
+
 ## 1.5.0
 
-- support dream 1.7 and dream-spec-helpers 1.2
+support dream 1.7 and dream-spec-helpers 1.2